Mostly improved comments.
Made horsep (horizontal with separators) behave then the number of objects is zero

A few more combinations of adjacent and nonadhjacent windows in a storey.

Removed the routine that randomly puts trees in fixed positions.  It had become dead code.

Trees are now placed across the entire street instead of in front of each building.

Reduced the number of buildings from 5 to 4 as a concession to machines with smaller screens.